DESKTOP: sync

+++ x11_core.c
@@ -49,7 +49,7 @@ int rect_dy = -6;
 int width;
 int height;
 
-uint32_t* core_init( int _width, int _height )
+uint8_t* core_init( int _width, int _height )
 {
   width  = _width;
   height = _height;
